requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# MMDetection Framework (optimizado para CUDA 12.8)
openmim>=0.3.9
mmcv>=2.0.0
mmdet>=3.0.0
mmengine>=0.8.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Metrics & Evaluation
pycocotools>=2.0.7
scikit-image>=0.21.0

# Data Processing
imageio>=2.31.0
tqdm>=4.66.0
shapely>=2.0.0

# Experiment Tracking & MLOps
mlflow>=3.6.0
wandb>=0.15.0
tensorboard>=2.14.0
cookiecutter>=2.6.0

# Utilities
pyyaml>=6.0
terminaltables>=3.1.0
colorama>=0.4.6

# Kaggle API (para descargar dataset)
kaggle>=1.5.16

# GPU Monitoring
nvidia-ml-py3>=7.352.0
gpustat>=1.1.1

# Configuration
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Image Processing
imagesize>=1.4.1
